requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
psycopg2-binary==2.9.11
bcrypt==4.1.2
redis==5.0.1
anthropic==0.18.1
openai==1.12.0
authlib==1.3.0
stripe==8.2.0
gunicorn==23.0.0